fix time_t logging

+++ content/llcache.c
@@ -3327,7 +3327,7 @@ void llcache_clean(bool purge)
 			NSLOG(llcache, DEBUG,
 			      "discarding fresh object len:%zd age:%ld (%p) %s",
 			      object->source_len,
-			      time(NULL) - object->last_used,
+			      (long)(time(NULL) - object->last_used),
 			      object,
 			      nsurl_access(object->url));
